+++ app/src/main/java/de/passwordvault/view/settings/activity_quality_gates/QualityGatesScreen.kt
@@ -7,6 +7,7 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
@@ -33,7 +34,7 @@ fun QualityGatesScreen(
                 title = {
                     Text(
                         text = stringResource(R.string.quality_gates),
-                        color = LocalPasswordVaultColors.current.text
+                        color = MaterialTheme.colorScheme.onSurface
                     )
                 },
                 navigationIcon = {
@@ -44,12 +45,12 @@ fun QualityGatesScreen(
                     ) {
                         Icon(
                             painter = painterResource(R.drawable.ic_back),
-                            tint = LocalPasswordVaultColors.current.text,
+                            tint = MaterialTheme.colorScheme.onSurface,
                             contentDescription = ""
                         )
                     }
                 },
-                colors = TopAppBarDefaults.topAppBarColors(containerColor = LocalPasswordVaultColors.current.backgroundAppBar)
+                colors = TopAppBarDefaults.topAppBarColors(containerColor = MaterialTheme.colorScheme.surfaceContainerHighest)
             )
         }
     ) { innerPadding ->
@@ -68,7 +69,7 @@ fun QualityGatesScreen(
                     onClick = {
                         //TODO
                     },
-                    modifier = Modifier.background(LocalPasswordVaultColors.current.background)
+                    modifier = Modifier.background(MaterialTheme.colorScheme.surface)
                 )
             }
         }